feat: 검색쿼리 구현(위치로직 제외)

+++ src/main/java/kr/co/suitcarrier/web/entity/Address.java
@@ -17,8 +17,8 @@ public class Address {
     private User user;
 
     @Column(name="longitude", nullable = false)
-    private Long longitude;
+    private double longitude;
 
     @Column(name="latitude", nullable = false)
-    private Long latitude;
+    private double latitude;
 }
